// generated_new/room/model/base-response-page-room-vo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { PageRoomVO } from './page-room-vo';

export interface BaseResponsePageRoomVO {
    'code'?: number;
    'data'?: PageRoomVO;
    'message'?: string;
}